+++ src/graphics/vec2/mod.rs
@@ -3,21 +3,21 @@
 use std::fmt::{self, Debug, Formatter};
 use zerocopy::{FromZeros, Immutable, IntoBytes, KnownLayout};
 
-#[repr(align(0x8), C)]
+#[repr(align(0x4), C)]
 #[derive(Clone, Copy, Default, FromZeros, Immutable, IntoBytes, KnownLayout, PartialEq, PartialOrd)]
-pub struct Vec2([f32; 0x2]);
+pub struct Vec2([f16; 0x2]);
 
 impl Vec2 {
 	#[inline(always)]
 	#[must_use]
-	pub const fn new(x: f32, y: f32) -> Self {
+	pub const fn new(x: f16, y: f16) -> Self {
 		let buf = [x, y];
 		Self(buf)
 	}
 
 	#[inline(always)]
 	#[must_use]
-	pub const fn get(self) -> (f32, f32) {
+	pub const fn get(self) -> (f16, f16) {
 		let [x, y] = self.0;
 		(x, y)
 	}
@@ -30,9 +30,9 @@ impl Debug for Vec2 {
 	}
 }
 
-impl From<(f32, f32)> for Vec2 {
+impl From<(f16, f16)> for Vec2 {
 	#[inline(always)]
-	fn from((x, y): (f32, f32)) -> Self {
+	fn from((x, y): (f16, f16)) -> Self {
 		Self::new(x, y)
 	}
 }
